chore: extract biobank specific actions
